Pulled logic out of SubscriptionDetailController
